Add executable jar

+++ Graph/Graph/src/com/gmail/at/ivanehreshi/graph/EdgeTo.java
@@ -4,7 +4,7 @@ public class EdgeTo {
 	public int to;
 	public double w = 0;
 	
-	public EdgeTo(int to, int w){
+	public EdgeTo(int to, double w){
 		this.to = to;
 		this.w = w;
 	}
